dataflow,sinks: move sink rendering code to sink implementations
The two were closely coupled before, some of the things the rendering
code would enforce, such as keying by the sink id, or consolidating
where depended upon by the actual sink code. Now all the code for a sink
is consolidated in one place and the generic sink rendering code is free
of sink specifics.
