+++ src/dataflow/src/sink/mod.rs
@@ -10,7 +10,3 @@
 mod avro_ocf;
 mod kafka;
 mod tail;
-
-pub use avro_ocf::avro_ocf;
-pub use kafka::kafka;
-pub use tail::tail;
